Lay the foundation for subtree switching
Left arrow key and 'h' both attempt to switch the tree to the
left. This exposed a problem with the server implementation in which
messages that don't have UUIDs could be broadcast. Fixing this was
surprisingly painful, as the server's self-replying mechanism makes
bad assumptions about how the ID numbers will be assigned.

The server's handling of messages (validation in general) needs to
be improved. The current implementation is full of hacks.